requirements.txt
alabaster==0.7.10
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.24.0
Babel==2.4.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
colorama==0.3.7
coverage==4.3.4
cryptography==2.2.2
dj-database-url==0.4.2
Django==2.0.4
django-filter==1.1.0
django-model-utils==3.1.1
django-phonenumber-field==1.3.0
django-storages==1.6.6
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
Faker==0.7.3
gunicorn==19.7.1
idna==2.6
imagesize==0.7.1
Jinja2==2.10
livereload==2.5.1
MarkupSafe==1.0
mixer==5.6.6
mock==2.0.0
packaging==16.8
paramiko==2.4.1
pathtools==0.1.2
pbr==3.0.0
phonenumberslite==8.4.1
port-for==0.3.1
psycopg2==2.7.4
py==1.4.33
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pyparsing==2.2.0
pytest==3.0.7
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-watch==4.1.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
six==1.10.0
snowballstemmer==1.2.1
sphinxcontrib-websupport==1.0.1
tornado==4.5.2
urllib3==1.22
watchdog==0.8.3
whitenoise==3.3.0
Markdown==2.6.11
pprint==0.1
simplejson==3.14.0
sqlparse==0.2.4